Sprint 16 MVVM 1.03
start moving to controller
