+++ app/src/main/java/com/practicum/playlistmaker/search/ui/SearchActivity.kt
@@ -39,36 +39,28 @@ class SearchActivity : ComponentActivity() {
     companion object {
         const val USERTEXT =
             "USER_INPUT"   //константа-ключ для поиска в Bundle сохраненного состояния
-        const val NOTHING_FOUND = "1"
-        const val SOMETHING_WENT_WRONG = "2"
-        private const val SEARCH_DEBOUNCE_DELAY = 2000L
     }
-    private lateinit var searchViewModel: SearchViewModel
-    private val searchInteractor = Creator.provideSearchInteractor()
+//    private lateinit var searchViewModel: SearchViewModel
 
     private lateinit var editTextSearchActivity: EditText
     private lateinit var searchClearEdittextImageview: ImageView
     private lateinit var searchArrowBack: androidx.appcompat.widget.Toolbar
     private lateinit var recyclerViewSearch: RecyclerView
     private lateinit var recyclerViewListenedTracks: RecyclerView
-    private lateinit var placeholderMessage: TextView
-    private lateinit var placeholderImage: ImageView
-    private lateinit var placeholderButtonReload: AppCompatButton
-    private lateinit var layoutOfListenedTracks: LinearLayout
+
     private lateinit var sharedPrefs: SharedPreferences
     private lateinit var searchHistoryClearButton: AppCompatButton
     private lateinit var listener: SharedPreferences.OnSharedPreferenceChangeListener
-    private lateinit var progressBar: ProgressBar
 
-    private var userInputText: String = ""
+
+
     private var trackList = ArrayList<Track>()
     private var trackListAdapter = SearchAdapter(trackList)
     private var selectedTracks = ArrayList<Track>()
     private var selectedTracksAdapter =
         SearchAdapter(selectedTracks)      //адаптер для прослушанных треков
 
-    private val handler = Handler(Looper.getMainLooper())
-    private val searchRunnable = Runnable { search() }
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -76,8 +68,8 @@ class SearchActivity : ComponentActivity() {
         finderViewById()
 
 
-        /*создаем viewModel*/
-        searchViewModel = ViewModelProvider(this,getViewModelFactory(userInputText))[SearchViewModel::class.java]
+        /*создаем viewModel*//*
+        searchViewModel = ViewModelProvider(this,getViewModelFactory(userInputText))[SearchViewModel::class.java]*/
 
         recyclerViewSearch.adapter = trackListAdapter
         sharedPrefs = getSharedPreferences(SHARED_PREFS_SELECTED_TRACKS, MODE_PRIVATE)
@@ -207,98 +199,6 @@ class SearchActivity : ComponentActivity() {
         layoutOfListenedTracks.visibility = View.GONE
     }
 
-    private fun search() {
-        if (userInputText.isNotEmpty()) {
-
-            placeholderMessage.visibility = View.GONE
-            placeholderImage.visibility = View.GONE
-            placeholderButtonReload.visibility = View.GONE
-            layoutOfListenedTracks.visibility = View.GONE
-            progressBar.visibility = View.VISIBLE
-
-            searchInteractor.searchTracks(userInputText, object: SearchInteractor.SearchConsumer{
-                override fun consume(foundTracks: List<Track>){
-                    handler.post {
-                        progressBar.visibility = View.GONE
-                        trackList.clear()
-                        trackList.addAll(foundTracks)
-                        recyclerViewSearch.visibility = View.VISIBLE
-                        trackListAdapter.notifyDataSetChanged()
-
-                        if (trackList.isNotEmpty()) {
-                            showMessage("", "")
-                            placeholderMessage.visibility = View.GONE
-                            placeholderImage.visibility = View.GONE
-                            placeholderButtonReload.visibility = View.GONE
-                        } else {
-                            showMessage(getString(R.string.nothing_found), NOTHING_FOUND)
-                        }
-                    }
-                }
-            })
-
-/*            itunesService.search(userInputText)
-                .enqueue(object : Callback<TrackSearchResponse> {
-                    override fun onResponse(
-                        call: Call<TrackSearchResponse>, response: Response<TrackSearchResponse>
-                    ) {
-                        progressBar.visibility = View.GONE
-                        when (response.code()) {
-
-                            200 -> {        //success
-                                if (response.body()?.results?.isNotEmpty() == true) {
-                                    placeholderMessage.visibility = View.GONE
-                                    placeholderImage.visibility = View.GONE
-                                    placeholderButtonReload.visibility = View.GONE
-                                    trackListAdapter.setTracks(
-                                        response.body()?.results!!
-                                    )
-                                    showMessage("", "")
-                                } else {
-                                    showMessage(getString(R.string.nothing_found), NOTHING_FOUND)
-                                }
-                            }
-
-                            else -> {
-                                //error with server answer
-                                showMessage(
-                                    getString(R.string.something_went_wrong),
-                                    SOMETHING_WENT_WRONG
-                                )
-                            }
-                        }
-                    }
-
-                    override fun onFailure( //error without server answer
-                        call: Call<TrackSearchResponse>, t: Throwable
-                    ) {
-                        progressBar.visibility = View.GONE
-                        showMessage(getString(R.string.something_went_wrong), SOMETHING_WENT_WRONG)
-                    }
-                })*/
-        }
-    }
-
-    private fun searchDebounce() {
-        handler.removeCallbacks(searchRunnable)
-        handler.postDelayed(searchRunnable, SEARCH_DEBOUNCE_DELAY)
-    }
-
-    private fun showMessage(text: String, myErrorCode: String) {
-        if (text.isNotEmpty()) {
-            trackListAdapter.setTracks(trackList)
-            placeholderMessage.text = text
-            if (myErrorCode == NOTHING_FOUND) {
-                placeholderImage.setImageResource(R.drawable.placeholder_nothing_found)
-                placeholderButtonReload.visibility = View.GONE
-            } else if (myErrorCode == SOMETHING_WENT_WRONG) {
-                placeholderImage.setImageResource(R.drawable.placeholder_no_network)
-                placeholderButtonReload.visibility = View.VISIBLE
-            }
-            placeholderMessage.visibility = View.VISIBLE
-            placeholderImage.visibility = View.VISIBLE
-        }
-    }
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
